removed reference to lombok from functional test and moved tests to right package
