+++ src/functionalTests/java/uk/gov/hmcts/reform/finrem/functional/TestContextConfiguration.java
@@ -3,7 +3,6 @@ package uk.gov.hmcts.reform.finrem.functional;
 
 import feign.Feign;
 import feign.jackson.JacksonEncoder;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.cloud.openfeign.support.SpringMvcContract;
@@ -15,7 +14,6 @@ import uk.gov.hmcts.reform.authorisation.ServiceAuthorisationApi;
 import uk.gov.hmcts.reform.authorisation.generators.ServiceAuthTokenGenerator;
 
 
-@Slf4j
 @Configuration
 @ComponentScan("uk.gov.hmcts.reform.finrem.functional")
 @EnableFeignClients(basePackageClasses = ServiceAuthorisationApi.class)
