--- src/test/java/uk/gov/hmcts/reform/finrem/caseorchestration/e2etest/BaseE2ETest.java
@@ -1,37 +1,0 @@
-package uk.gov.hmcts.reform.finrem.caseorchestration.e2etest;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.junit.Before;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.annotation.PropertySource;
-import org.springframework.test.annotation.DirtiesContext;
-import org.springframework.test.web.servlet.MockMvc;
-import uk.gov.hmcts.reform.finrem.caseorchestration.BaseTest;
-import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CCDRequest;
-
-import java.io.IOException;
-import java.io.InputStream;
-
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@PropertySource(value = "classpath:application.properties")
-@AutoConfigureMockMvc
-@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-abstract class BaseE2ETest extends BaseTest {
-
-    @Autowired
-    protected ObjectMapper objectMapper;
-
-    @Autowired
-    protected MockMvc webClient;
-
-    protected CCDRequest request;
-
-    @Before
-    public void setUp() throws IOException {
-        try (InputStream resourceAsStream = getClass().getResourceAsStream("/fixtures/fee-lookup.json")) {
-            request = objectMapper.readValue(resourceAsStream, CCDRequest.class);
-        }
-    }
-}
